// src/core/http/generated/models/tenantUser.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */

export interface TenantUser {
  id: string;
  name: string;
  email: string;
  role?: string;
  active?: boolean;
  phone?: string;
  document?: string;
  created_at?: string;
  updated_at?: string;
}