// src/core/http/generated/models/createTenantUserPayload.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */

export interface CreateTenantUserPayload {
  name: string;
  email: string;
  password?: string;
  password_confirmation?: string;
  role?: string;
  active?: boolean;
  phone?: string;
  document?: string;
}